* move Collection<> definition from src/MSONSourcemap.h back to src/Blueprint.h * minor changes in #include to remove Blueprint.h dependenci on MSONSourcemap
